Migrate to Redux store
